+++ src/background/handlers/system.js
@@ -7,6 +7,18 @@ import {
 } from "../../common/index.js";
 
 const updateDefaultValues = async () => {
+    // execution
+    const isAutomaticEnabled = await Sync.get(OPTIONS.IS_AUTOMATIC_EXECUTION_ENABLED);
+    if (isAutomaticEnabled === undefined) {
+        await Sync.set(OPTIONS.IS_AUTOMATIC_EXECUTION_ENABLED, true);
+    }
+
+    const isLoggingEnabled = await Sync.get(OPTIONS.IS_EXECUTION_LOGGING_ENABLED);
+    if (isLoggingEnabled === undefined) {
+        await Sync.set(OPTIONS.IS_EXECUTION_LOGGING_ENABLED, true);
+    }
+
+    // autoclose
     const isAutocloseEnabled = await Sync.get(OPTIONS.IS_AUTOCLOSE_ENABLED);
     if (isAutocloseEnabled === undefined) {
         await Sync.set(OPTIONS.IS_AUTOCLOSE_ENABLED, true);
@@ -17,6 +29,7 @@ const updateDefaultValues = async () => {
         await Sync.set(OPTIONS.AUTOCLOSE_TIME, 1);
     }
 
+    // local data
     const targets = await Local.get(TARGETS.TARGETS);
     if (targets === undefined) {
         await Local.set(TARGETS.TARGETS, []);