+++ src/common/constants/options.js
@@ -1,2 +1,5 @@
 export const IS_AUTOCLOSE_ENABLED = "is-autoclose-enabled";
 export const AUTOCLOSE_TIME = "autoclose-time";
+
+export const IS_AUTOMATIC_EXECUTION_ENABLED = "is-automatic-execution-enabled";
+export const IS_EXECUTION_LOGGING_ENABLED = "is-execution-logging-enabled";