+++ src/common/index.js
@@ -1,5 +1,5 @@
 export * from "./storages/index.js";
 export * from "./models/index.js";
 export * from "./constants/index.js";
-export * from "./funcs/index.js";
+export * as funcs from "./funcs/index.js";
 export * as dom from "./dom/index.js";