+++ src/common/models/target.js
@@ -1,9 +1,10 @@
 export class Target {
-    constructor(id, name, pattern, selector, isActive) {
+    constructor(id, name, pattern, selector, isActive, isAuto) {
         this.id = id;
         this.name = name;
         this.pattern = pattern;
         this.selector = selector;
         this.isActive = isActive;
+        this.isAuto = isAuto;
     }
 }