+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-04-29
+
+### Added
+
+- Automatic execution option and target property
+- Execution logging
+
 ## [0.3.0] - 2023-04-12
 
 ### Added
@@ -26,10 +33,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add multiple targets: name, URL pattern and DOM element selector
 - Activate extension (invoke popup) and press Enter to try to click according to all matched targets and stop after first that have only one corresponding DOM element
-- Use regular expression for URL mathing
+- Use regular expression for URL matching
 - Configure popup autoclosing and autoclosing time
 
-[Unreleased]: https://github.com/vikian050194/click/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/vikian050194/click/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/vikian050194/click/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/vikian050194/click/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/vikian050194/click/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/vikian050194/click/releases/tag/v0.1.0